api: remove an error from discoverers constructors
The API designed to make chains of decorators. In case of the
Discoverer object it will look like:

```
	// The object construction.
	disc := discoverer.NewRetry(
		discoverer.NewFilter(
			discoverer.NewEtcd(etcd, "foo"),
			filter.FooFilter{}, filter.BarFilter{},
		),
		10,
		logger,
	)

	// Usage.
	instances, err := disc.Discover(ctx)
```

This doesn't work very well for cases where constructors return
an error as a second return value. The code turns out to be complex
and multi-stage. So let's assume that the library user creates the
object correctly. And if not, then he will see an error the first
time he uses the object.

Part of #5