test: explicit fail for too old Tarantool
Library works only with Tarantool 3 due to WATCH_ONCE feature use [1].
If one tries to run tests with Tarantool 2.11 or older, some asserts
will fail, but the reason will be rather unclear. (This is what had
happened to me.) This patch adds explicit version check for tests and
examples starting a Tarantool instance.

1. https://github.com/tarantool/tarantool/pull/8741